filters: implement filters utility class that wraps extensions, targets and ignores
This filters class makes it easier to deal with the customization for the filtering during the path selection process. Especially because we support both global filters and dependency-specific filters, which means we have to manipulate these arrays to merge them, etc...

As part of this change we also refine and simplify the `Preset` interface, since now it only has to implement two methods for getting the filters:

  - `GetFilters() *Filters`
  - `GetFiltersForDependency(*Dependency) *Filters`

That should make things simpler to deal with at that level too.