preset: support customising the repository cache path
The default preset now attempts to use a user HOME path.
If there is no user HOME, then fall back to OS tempdir.

Not sure if a warning should be logged when this happens. I've experienced
cache corruption issues, in which git enters a broken state and cannot checkout
branches or change the worktree.