git: reliable install and updates due to reliance on a forced reset (#24)
This patch cleans up the git operations file, now it matches what we
originally did on the `vendor-rs` tool. Now we rely on fetching and
resetting the work-tree. This ensures local changes are discarded and
also lays the groundwork so in the future I can bring back the support
for any kind of reference name in the dependencies of the spec, which
now are constrained to being only branches. In the future, they will
be able to be any kind of reference (tag, head, branch, commit...)