orchestrator: move command logic into orchestrator struct
This simplifies the cobra CLI Run functions
As part of this refactor, updated all functions in the codebase
to wrap the errors they return. This has simplified as well
the error handling, so the error bubbles up and is only logged
once it reaches the Run() of the cobra command.