internal/installer: vendor dependencies in parallel
Dependencies are now vendored in parallel, the speedup is quite
noticeable, specially for update operations, since most of the time
is spent fetching the remote, which is I/O bound.

Since the operations are no longer sequential, had to tune the output
logging so that the user understands what is being vendored and locked.