internal/importer: simplify importer logic
Splits selector in two methods 1) SelectPath and 2) SelectDir
Now importer will call each method depending on the kind of entry that WalkDir reports.
This should make things slightly more efficient, since we no longer need to compute
whether a directory must be entered when we are processing a filepat, and viceversa;
we should not be computing wether to select a filepath when processing a directory.