internal/installer: fetch during install when the git reset fails
This patch adds conditional fetching of the remote when the git reset
operation fails. This ensures we try to fetch the remote just in case the install
operation is being executed on a reference that is not available on the current
cache.

This prevents issues when the remote has advanced, perhaps by other persons or because
you pushed from a different machine, but your local cache gets behind. You would not
usually run an install operation, but if you ever did, you would see an error because
the reference may not be available.